requirements.txt
runpod
torch
diffusers
transformers
accelerate
pillow
scipy